requirements.txt
Django==2.2.4
djangorestframework==3.10.2
pytrends==4.6.0
django-cors-headers==3.1.0
requests==2.22.0